+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [0.2.0] - 2026-09-25
 
 ### Added
 - `subscribers.purged` and `subscribers.purge_failed` webhook event types, in